+++ silo-core/test/foundry/gas/Deposit1st.gas.sol
@@ -15,7 +15,7 @@ contract Deposit1stGasTest is Gas, Test {
         _gasTestsInit();
     }
 
-    // forge test -vv --ffi --mt test_gas_firstDeposit
+    // FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_gas_firstDeposit
     //  194207 - when __accrueInterest returns 2 configs
     // -188200  when __accrueInterest returns config and we pul configs in lib
     function test_gas_firstDeposit() public {